Image pull secrets should be defined as kubernetes manifests.
The feature for automatic creation of an image pull secret for
docker.io has been removed.

Image pull secrets should now be defined as a kubernetes manifest
included in the kustomization.  The instructions in the README
have been updated to show how to add an image pull secret
for Dockerhub.

Change-type: minor
Signed-off-by: Carlo Miguel F. Cruz <carloc@balena.io>